# openapi.yaml
openapi: 3.0.3
info:
  title: Point of Sale (POS) System API
  version: 1.0.0
  description: >
    Comprehensive REST API documentation for POS System Backend featuring multi-branch management,
    real-time analytics, inventory control, cashier shift management, refund processing, and payment integration.
  contact:
    name: POS System Support
    email: support@pos-system.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
security:
  - BearerAuth: []
paths:
...
  /api/user/profile:
    get:
      tags:
        - User Profile Management
      summary: Get Authenticated User Profile
      description: Retrieves the user profile and role of the currently authenticated user from their JWT Bearer token.
      responses:
        '200':
          description: Successfully retrieved user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '401':
          description: Unauthorized / Invalid JWT token
  /api/user/{id}:
    get:
      tags:
        - User Profile Management
      summary: Get User Profile by ID
      description: Retrieves user details by user ID (restricted to self, Store Admin, Store Manager, or Branch Manager).
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the target user
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved user details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '403':
          description: Access denied

  # ---------------------------------------------------------------------------
  # Store Management
  # ---------------------------------------------------------------------------
  /api/store:
    post:
      tags:
        - Store Management
      summary: Create Store
      description: Registers a new store for the authenticated store admin user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreDTO'
      responses:
        '201':
          description: Store successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreDTO'
    get:
      tags:
        - Store Management
      summary: Get All Stores
      description: Retrieves a list of all registered stores in the system.
      responses:
        '200':
          description: Successfully retrieved stores list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StoreDTO'
  /api/store/{id}:
    get:
      tags:
        - Store Management
      summary: Get Store by ID
      description: Retrieves store details by store ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved store details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreDTO'
    put:
      tags:
        - Store Management
      summary: Update Store
      description: Updates an existing store's profile by store ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreDTO'
      responses:
        '200':
          description: Store successfully updated
    delete:
      tags:
        - Store Management
      summary: Delete Store
      description: Deletes a store by store ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Store successfully deleted
...
  /api/store/{id}/moderate:
    put:
      tags:
        - Store Management
      summary: Moderate Store Status
      description: Allows Super Admin to update store status (e.g. ACTIVE, INACTIVE, SUSPENDED).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          required: true
          schema:
            type: string
            enum: [ACTIVE, INACTIVE, SUSPENDED]
      responses:
        '200':
          description: Store status successfully updated

  # ---------------------------------------------------------------------------
  # Branch Management
  # ---------------------------------------------------------------------------
...
  /api/branch/store/{storeId}:
    get:
      tags:
        - Branch Management
      summary: Get All Branches by Store ID
      description: Retrieves a list of all branches belonging to a specific store.
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved branch list

  # ---------------------------------------------------------------------------
  # Employee Management
  # ---------------------------------------------------------------------------
  /api/employee/store/{storeId}:
    post:
      tags:
        - Employee Management
      summary: Create Store Employee
      description: Creates a new employee (e.g. Store Admin/Manager) associated with a store.
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        '201':
          description: Store employee successfully created
    get:
      tags:
        - Employee Management
      summary: Find All Store Employees
      description: Retrieves all employees for a store, optionally filtered by user role.
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userRole
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved store employees list
...
  /api/employee/{employeeId}:
    put:
      tags:
        - Employee Management
      summary: Update Employee
      description: Updates employee details (name, email, role, phone) by employee ID.
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        '200':
          description: Employee successfully updated
    delete:
      tags:
        - Employee Management
      summary: Delete Employee
      description: Removes an employee by employee ID.
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Employee successfully deleted

  # ---------------------------------------------------------------------------
  # Category Management
  # ---------------------------------------------------------------------------
...
  /api/categories/store/{storeId}:
    get:
      tags:
        - Category Management
      summary: Get Categories by Store ID
      description: Retrieves all product categories belonging to a specific store.
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved categories list
...
  /api/products/store/{storeId}:
    get:
      tags:
        - Product Management
      summary: Get Products by Store ID
      description: Retrieves all products belonging to a specific store catalog.
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved store products catalog
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT Bearer token obtained from `/auth/login` or `/auth/signup`
  schemas:
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [ROLE_ADMIN, ROLE_STORE_ADMIN, ROLE_STORE_MANAGER, ROLE_BRANCH_MANAGER, ROLE_BRANCH_CASHIER]
        storeId:
          type: integer
          format: int64
        branchId:
          type: integer
          format: int64
...
    StoreDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        brand:
          type: string
        description:
          type: string
        storeAdminId:
          type: integer
          format: int64
        status:
          type: string
          enum: [ACTIVE, INACTIVE, SUSPENDED]
...